refactor: move parse_args() to infrastructure helper
