+++ tests/modules/variables/variables_support_test.py
@@ -4,23 +4,7 @@ import pytest
 from chk.infrastructure.containers import App
 from chk.infrastructure.file_loader import FileContext
 from chk.infrastructure.helper import data_set
-from chk.modules.variables.support import parse_args, VariableMixin, replace_values
-
-
-class TestParseArgs:
-    def test_parse_args_pass_for_unique(self):
-        argv_s = ["Var1=Val1", "Var2=Val2", "Var3=Val3", "Var=Val"]
-        response = parse_args(argv_s)
-
-        assert isinstance(response, dict)
-        assert len(response) == 4
-
-    def test_parse_args_pass_for_override(self):
-        argv_s = ["Var1=Val1", "Var2=Val2", "Var3=Val3", "Var1=Val4"]
-        response = parse_args(argv_s)
-
-        assert isinstance(response, dict)
-        assert len(response) == 3
+from chk.modules.variables.support import VariableMixin, replace_values
 
 
 class HavingVariables(VariableMixin):
